wiki_core/requirements.txt
Django==2.1.5
djangorestframework==3.9.0
psycopg2-binary==2.7.6.1
pytz==2018.9
